Make next module step
